backend/requirements.txt
# Core FastAPI and web server dependencies
fastapi==0.115.6
uvicorn[standard]==0.32.1
slowapi==0.1.9

# PDF processing
pdfplumber==0.11.4

# AI APIs
google-genai>=1.0.0
gtts==2.5.1

# Database & Auth
supabase>=2.4.0
sqlalchemy==2.0.25
alembic==1.13.1
python-jose[cryptography]==3.3.0
psycopg2-binary>=2.9.9

# File upload handling
python-multipart==0.0.7

# Environment and configuration
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.8.1

# HTTP clients and async utilities
httpx>=0.24
aiofiles==23.2.1

# Vercel Blob storage
vercel-blob==0.2.0

# Image processing (Python 3.13 compatible version)
# Pillow==10.3.0  # Temporarily disabled due to Python 3.13 build issues

# Development and testing (temporarily removed for Python 3.13 compatibility)
# pytest==7.4.4
# pytest-asyncio==0.21.1

# Note: Manim and video generation dependencies are installed in the agent_env
# virtual environment (see agent_requirements.txt), not in the main backend environment
